refactor(br41a): extract local-tool protocol types + portable permission schema into cowork-bridge; chrome consumes them
