+++ packages/cowork-bridge/src/permissions/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Portable tool-permission schema + matching — barrel export. Persistence and
+ * backend sync stay host-side.
+ */
+
+export * from './types.js';
+export * from './matching.js';
