+++ packages/cowork-bridge/src/tools/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Local-tool protocol — barrel export. Types only; execution is host-supplied.
+ */
+
+export * from './types.js';
