+++ ui/chrome-ext/tool-permissions.ts
@@ -1,28 +1,30 @@
 import { getValidAccessToken } from './extension-auth';
 import { loadExtensionConfig } from './extension-config';
+import {
+    getOriginPatternScore,
+    getToolPatternScore,
+    matchesOriginPattern,
+    matchesToolPattern,
+    normalizeEntry,
+    normalizeOriginPattern,
+    normalizePolicy,
+    normalizeRuntimeOrigin,
+    normalizeToolNamePattern,
+    policyKey,
+    resolveMatchingPolicy as resolveMatchingPolicyFromEntries,
+    type ToolPermissionDecision,
+    type ToolPermissionEntry,
+    type ToolPermissionPolicy,
+    type ToolPermissionRequest,
+} from '@sentropic/cowork-bridge/permissions';
 
-export type ToolPermissionPolicy = 'allow' | 'deny';
-export type ToolPermissionDecision =
-    | 'allow_once'
-    | 'deny_once'
-    | 'allow_always'
-    | 'deny_always';
-
-export type ToolPermissionEntry = {
-    toolName: string;
-    origin: string;
-    policy: ToolPermissionPolicy;
-    updatedAt: string;
-};
-
-export type ToolPermissionRequest = {
-    requestId: string;
-    toolName: string;
-    origin: string;
-    tabId?: number;
-    tabUrl?: string;
-    tabTitle?: string;
-    details?: Record<string, unknown>;
+// Re-export the portable permission schema under the historical names so
+// existing consumers keep importing them from this module.
+export type {
+    ToolPermissionDecision,
+    ToolPermissionEntry,
+    ToolPermissionPolicy,
+    ToolPermissionRequest,
 };
 
 type PermissionCheckInput = {
@@ -71,204 +73,11 @@ let loaded = false;
 let loadPromise: Promise<void> | null = null;
 let lastSyncAt = 0;
 
-const policyKey = (toolName: string, origin: string) => `${toolName}::${origin}`;
-
-const TOOL_PATTERN_REGEX = /^[a-z0-9:_*-]{1,96}$/i;
-const HOSTNAME_LABEL_REGEX = /^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$/i;
-const IPV4_REGEX =
-    /^(?:25[0-5]|2[0-4]\d|1?\d?\d)(?:\.(?:25[0-5]|2[0-4]\d|1?\d?\d)){3}$/;
-
-const normalizeToolNamePattern = (raw: string): string | null => {
-    const value = raw.trim().toLowerCase();
-    if (!value) return null;
-    if (!TOOL_PATTERN_REGEX.test(value)) return null;
-    if (value.includes('**')) return null;
-    return value;
-};
-
-const isValidHostname = (host: string): boolean => {
-    const value = host.trim().toLowerCase();
-    if (!value) return false;
-    if (value === 'localhost') return true;
-    if (IPV4_REGEX.test(value)) return true;
-    const labels = value.split('.');
-    if (labels.length < 2) return false;
-    return labels.every((label) => HOSTNAME_LABEL_REGEX.test(label));
-};
-
-const normalizeRuntimeOrigin = (raw: string): string | null => {
-    const value = raw.trim();
-    if (!value) return null;
-    try {
-        const parsed = new URL(value);
-        if (parsed.protocol !== 'http:' && parsed.protocol !== 'https:') return null;
-        const hostname = parsed.hostname.toLowerCase();
-        const port = parsed.port ? `:${parsed.port}` : '';
-        return `${parsed.protocol}//${hostname}${port}`;
-    } catch {
-        return null;
-    }
-};
-
-const normalizeOriginPattern = (raw: string): string | null => {
-    const value = raw.trim().toLowerCase();
-    if (!value) return null;
-    if (value === '*') return '*';
-
-    // all hosts with explicit scheme
-    const schemeAnyHostMatch = value.match(/^(https?:)\/\/\*$/);
-    if (schemeAnyHostMatch) {
-        return `${schemeAnyHostMatch[1]}//*`;
-    }
-
-    // host wildcard (all schemes)
-    if (value.startsWith('*.')) {
-        const suffix = value.slice(2);
-        if (!isValidHostname(suffix)) return null;
-        return `*.${suffix}`;
-    }
-
-    // host wildcard with explicit scheme (http/https only)
-    const wildcardSchemeMatch = value.match(/^(https?:)\/\/\*\.(.+)$/);
-    if (wildcardSchemeMatch) {
-        const scheme = wildcardSchemeMatch[1];
-        const suffix = wildcardSchemeMatch[2];
-        if (!isValidHostname(suffix)) return null;
-        return `${scheme}//*.${suffix}`;
-    }
-
-    // exact host pattern (all schemes)
-    if (isValidHostname(value)) {
-        return value;
-    }
-
-    // exact URL origin (with scheme)
-    return normalizeRuntimeOrigin(value);
-};
-
-const wildcardPatternToRegExp = (pattern: string): RegExp => {
-    const escaped = pattern.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-    return new RegExp(`^${escaped.replace(/\\\*/g, '.*')}$`);
-};
-
-const matchesToolPattern = (pattern: string, toolName: string): boolean => {
-    if (pattern === '*') return true;
-    if (pattern === toolName) return true;
-    // Backward-compat: legacy `tab_action` matches `tab_action:*`.
-    if (!pattern.includes('*') && !pattern.includes(':')) {
-        return toolName.startsWith(`${pattern}:`);
-    }
-    if (!pattern.includes('*')) return false;
-    return wildcardPatternToRegExp(pattern).test(toolName);
-};
-
-const matchesOriginPattern = (
-    pattern: string,
-    runtimeOrigin: string,
-): boolean => {
-    if (pattern === '*') return true;
-    const parsed = new URL(runtimeOrigin);
-    const hostname = parsed.hostname.toLowerCase();
-
-    const schemeAnyHostMatch = pattern.match(/^(https?:)\/\/\*$/);
-    if (schemeAnyHostMatch) {
-        return parsed.protocol === schemeAnyHostMatch[1];
-    }
-
-    if (pattern.startsWith('*.')) {
-        const suffix = pattern.slice(2);
-        return hostname === suffix || hostname.endsWith(`.${suffix}`);
-    }
-
-    const wildcardSchemeMatch = pattern.match(/^(https?:)\/\/\*\.(.+)$/);
-    if (wildcardSchemeMatch) {
-        const requiredScheme = wildcardSchemeMatch[1];
-        const suffix = wildcardSchemeMatch[2];
-        if (parsed.protocol !== requiredScheme) return false;
-        return hostname === suffix || hostname.endsWith(`.${suffix}`);
-    }
-
-    if (pattern.includes('://')) {
-        return pattern === runtimeOrigin;
-    }
-
-    // exact hostname rule (all schemes)
-    return hostname === pattern;
-};
-
-const getToolPatternScore = (pattern: string): number => {
-    if (pattern === '*') return 0;
-    const literalChars = pattern.replace(/\*/g, '').length;
-    if (!pattern.includes('*')) {
-        return (pattern.includes(':') ? 3000 : 2500) + literalChars;
-    }
-    return 1500 + literalChars;
-};
-
-const getOriginPatternScore = (pattern: string): number => {
-    if (pattern === '*') return 0;
-    if (pattern.includes('://') && !pattern.includes('*')) return 4000 + pattern.length;
-    if (pattern.includes('://*.')) return 3000 + pattern.length;
-    if (pattern.startsWith('*.')) return 2000 + pattern.length;
-    if (pattern.endsWith('://*')) return 1000 + pattern.length;
-    return 3500 + pattern.length; // exact hostname
-};
-
 const resolveMatchingPolicy = (
     toolName: string,
     runtimeOrigin: string,
-): ToolPermissionEntry | null => {
-    const matches = Array.from(policiesByKey.values()).filter(
-        (entry) =>
-            matchesToolPattern(entry.toolName, toolName) &&
-            matchesOriginPattern(entry.origin, runtimeOrigin),
-    );
-    if (matches.length === 0) return null;
-    matches.sort((a, b) => {
-        const scoreA =
-            getOriginPatternScore(a.origin) * 10_000 +
-            getToolPatternScore(a.toolName);
-        const scoreB =
-            getOriginPatternScore(b.origin) * 10_000 +
-            getToolPatternScore(b.toolName);
-        if (scoreA !== scoreB) return scoreB - scoreA;
-        const updatedA = Date.parse(a.updatedAt);
-        const updatedB = Date.parse(b.updatedAt);
-        if (Number.isFinite(updatedA) && Number.isFinite(updatedB)) {
-            return updatedB - updatedA;
-        }
-        return 0;
-    });
-    return matches[0] ?? null;
-};
-
-const normalizePolicy = (value: string): ToolPermissionPolicy | null => {
-    if (value === 'allow' || value === 'deny') return value;
-    return null;
-};
-
-const normalizeEntry = (raw: {
-    toolName?: string;
-    origin?: string;
-    policy?: string;
-    updatedAt?: string;
-}): ToolPermissionEntry | null => {
-    const toolName = normalizeToolNamePattern(String(raw.toolName ?? ''));
-    const origin = normalizeOriginPattern(String(raw.origin ?? ''));
-    const policy = normalizePolicy(String(raw.policy ?? ''));
-    const updatedAtRaw = String(raw.updatedAt ?? '').trim();
-    const updatedAt =
-        updatedAtRaw && Number.isFinite(Date.parse(updatedAtRaw))
-            ? new Date(updatedAtRaw).toISOString()
-            : new Date().toISOString();
-    if (!toolName || !origin || !policy) return null;
-    return {
-        toolName,
-        origin,
-        policy,
-        updatedAt,
-    };
-};
+): ToolPermissionEntry | null =>
+    resolveMatchingPolicyFromEntries(policiesByKey.values(), toolName, runtimeOrigin);
 
 const parseStoredPolicies = (payload: unknown): ToolPermissionEntry[] => {
     if (!Array.isArray(payload)) return [];
